♻️Refactor: 이미지 업로드 S3 폴더로 변경
