+++ src/main/java/com/umc7th/a1grade/domain/question/controller/QuestionController.java
@@ -2,7 +2,6 @@ package com.umc7th.a1grade.domain.question.controller;
 
 import jakarta.validation.Valid;
 
-import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,9 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 import com.umc7th.a1grade.domain.question.dto.QuestionRequestDTO;
 import com.umc7th.a1grade.domain.question.dto.QuestionResponseDTO;
@@ -20,8 +17,6 @@ import com.umc7th.a1grade.domain.question.exception.status.QuestionErrorStatus;
 import com.umc7th.a1grade.domain.question.service.QuestionService;
 import com.umc7th.a1grade.global.annotation.ApiErrorCodeExample;
 import com.umc7th.a1grade.global.apiPayload.ApiResponse;
-import com.umc7th.a1grade.global.s3.PathName;
-import com.umc7th.a1grade.global.s3.exception.status.S3ErrorStatus;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -109,15 +104,5 @@ public class QuestionController {
       @Parameter(description = "질문 번호", example = "1") @PathVariable Long questionId) {
     QuestionResponseDTO.GetAnswerDTO answer = questionService.getAnswer(questionId);
     return ApiResponse.onSuccess(answer);
-  }
-
-  @Operation(summary = "이미지 업로드 API", description = "이미지를 업로드하고 URL을 리턴받는 API")
-  @ApiErrorCodeExample(S3ErrorStatus.class)
-  @PostMapping(value = "/img-upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-  public ApiResponse<QuestionResponseDTO.ImgUrlDTO> uploadImage(
-      @RequestParam PathName pathName, MultipartFile file) {
-
-    QuestionResponseDTO.ImgUrlDTO response = questionService.ImgUpload(pathName, file);
-    return ApiResponse.onSuccess(response);
   }
 }
