+++ src/main/java/com/umc7th/a1grade/domain/question/converter/QuestionConverter.java
@@ -88,8 +88,4 @@ public class QuestionConverter {
         .questions(falseQuestionDTOS)
         .build();
   }
-
-  public QuestionResponseDTO.ImgUrlDTO toImgUrlDTO(String imgUrl) {
-    return QuestionResponseDTO.ImgUrlDTO.builder().imageUrl(imgUrl).build();
-  }
 }
